requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-cors-headers==4.1.0
django-enumfields==2.1.1
django-recaptcha==3.0.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-rest-swagger==2.2.0
django-swagger-ui==0.1.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-enum-field==0.9.3
drf-spectacular==0.26.4
drf-yasg==1.21.10
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
Pygments==2.15.1
PyJWT==2.7.0
pyotp==2.9.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
uuid==1.30
xlwt==1.3.0
